lira/requirements.txt
# requirements.txt

python-telegram-bot==20.3
SQLAlchemy==1.4.46
python-dotenv==1.0.0
pytest==7.3.1
